+++ test/utest/test_query.py
@@ -44,16 +44,16 @@ class QueryTests(unittest.TestCase):
 
     def test_should_return_expected_host(self):
         """Simulate query to return session host endpoint URL."""
-        self.engine._connection = mock.PropertyMock()
-        self.engine._connection.host = 'https://dynamodb.MY-REGION.amazonaws.com'
+        self.engine.connection = mock.PropertyMock()
+        self.engine.connection.host = 'https://dynamodb.MY-REGION.amazonaws.com'
         self.query._cache.switch.return_value = self.engine
         response = self.query.dynamodb_host(self.label)
         self.assertEqual(response, 'https://dynamodb.MY-REGION.amazonaws.com')
 
     def test_should_return_expected_region(self):
         """Simulate query to return session region."""
-        self.engine._connection = mock.PropertyMock()
-        self.engine._connection.region = 'MY-REGION'
+        self.engine.connection = mock.PropertyMock()
+        self.engine.connection.region = 'MY-REGION'
         self.query._cache.switch.return_value = self.engine
         response = self.query.dynamodb_region(self.label)
         self.assertEqual(response, 'MY-REGION')
@@ -61,12 +61,12 @@ class QueryTests(unittest.TestCase):
     def test_should_return_table_list(self):
         """Simulate query to return table list."""
         response = mock.create_autospec(ResultSet)
-        self.engine._connection = mock.PropertyMock()
-        self.engine._connection.list_tables = mock.Mock(return_value=response)
+        self.engine.connection = mock.PropertyMock()
+        self.engine.connection.list_tables = mock.Mock(return_value=response)
         self.query._cache.switch.return_value = self.engine
         self.query.list_dynamodb_tables(self.label)
         self.query._cache.switch.assert_called_with(self.label)
-        self.engine._connection.list_tables.assert_called()
+        self.engine.connection.list_tables.assert_called()
         self.query._builtin.log.assert_called_with("List tables response:\n%s" %
                                                    "[]", 'DEBUG')