+++ CHANGELOG.rst
@@ -1,3 +1,30 @@
+0.2.0 (2015.08.02)
+==================
+
+* Use boto3
+* Add unit tests
+* Add dependency to boto3
+* Update documentation
+
+0.1.2 (2015.08.01)
+==================
+
+* Bugfixes
+
+0.1.1 (2015.08.01)
+==================
+
+* Add `DynamoDB Host`, `DynamoDB Region`, and `List DynamoDB Tables` keywords.
+* Add unit and acceptance tests
+* Update make file
+
+0.1.0 (2015.08.01)
+==================
+
+* Add support for config and credentials file
+* Add unit and acceptance tests
+* Update readme file
+
 0.0.15 (2015.07.31)
 ===================